asciidoctor-editor-plugin/src/main/java/de/jcup/asciidoctoreditor/asciidoc/Sha256StringEncoder.java
package de.jcup.asciidoctoreditor.asciidoc;

import java.nio.charset.StandardCharsets;
import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;

public class Sha256StringEncoder implements StringEncoder {

    private final static MessageDigest SHA_256_DIGEST = createDigest();

    public String encode(String originalString) {
        if (originalString == null) {
            return null;
        }
        byte[] encodedhash = SHA_256_DIGEST.digest(originalString.getBytes(StandardCharsets.UTF_8));

        return bytesToHex(encodedhash);
    }

    private static String bytesToHex(byte[] hash) {
        StringBuilder hexString = new StringBuilder(2 * hash.length);
        for (int i = 0; i < hash.length; i++) {
            String hex = Integer.toHexString(0xff & hash[i]);
            if (hex.length() == 1) {
                hexString.append('0');
            }
            hexString.append(hex);
        }
        return hexString.toString();
    }

    private static MessageDigest createDigest() {
        try {
            return MessageDigest.getInstance("SHA-256");
        } catch (NoSuchAlgorithmException e) {
            throw new IllegalStateException("SHA-256 must be available!");
        }
    }
}
